// YOUCOM.ReserVook/src/app/features/web/master/pointconvert/model/pointconvert.model.ts
export class PointConvertInfo {
  companyNo: string;
  scCd: string;
  scSiteCd: string;
  scPntsDiscntNm: string;
  denominationCode: string;
  updateCnt: number;
  programId: string;
  createClerk: string;
  createMachineNo: string;
  createMachine: string;
  createDatetime: string;
  updateClerk: string;
  updateMachineNo: string;
  updateMachine: string;
  updateDatetime: string;
  status: string;

  cdName: string;
  siteCdName: string;
  pntsDiscntName: string;
  denomination: string;
}